+++ src/components/pay-change-distribution.tsx
@@ -8,7 +8,6 @@ import {
   TableHeader,
   TableRow,
 } from '@/components/ui/table'
-import { MIN_JOBS_SHOWN } from '@/lib/department-jobs'
 import { formatCount } from '@/lib/format'
 import {
   type ChangeDistribution,
@@ -17,6 +16,7 @@ import {
   pairLabel,
 } from '@/lib/pay-changes'
 import { stackedCounts } from '@/lib/trend-groups'
+import { MIN_JOBS_SHOWN } from '@/lib/trends'
 
 const NUMBER_CELL = 'text-right tabular-nums'